Feature/mttl 1810 (#22)
* Test new DynamoDb testing package.

* Updated packages

* Updated DynamoDb test package version

* Renamed DynamoDbCollection

* Updated hackney nuget packages

* Updated Hackney.Core.Testing.DynamoDb package
