+++ AssetInformationListener.Tests/E2ETests/Fixtures/AssetFixture.cs
@@ -1,5 +1,5 @@
-using Amazon.DynamoDBv2.DataModel;
 using AutoFixture;
+using Hackney.Core.Testing.DynamoDb;
 using Hackney.Shared.Asset.Infrastructure;
 using Hackney.Shared.Tenure.Boundary.Response;
 using System;
@@ -9,16 +9,15 @@ namespace AssetInformationListener.Tests.E2ETests.Fixtures
     public class AssetFixture : IDisposable
     {
         private readonly Fixture _fixture = new Fixture();
-        private const string DateFormat = "yyyy-MM-ddTHH\\:mm\\:ss.fffffffZ";
 
-        private readonly IDynamoDBContext _dbContext;
+        private readonly IDynamoDbFixture _dbFixture;
 
         public AssetDb DbAsset { get; private set; }
         public Guid AssetDbId { get; private set; }
 
-        public AssetFixture(IDynamoDBContext dbContext)
+        public AssetFixture(IDynamoDbFixture dbFixture)
         {
-            _dbContext = dbContext;
+            _dbFixture = dbFixture;
         }
 
         public void Dispose()
@@ -32,9 +31,6 @@ namespace AssetInformationListener.Tests.E2ETests.Fixtures
         {
             if (disposing && !_disposed)
             {
-                if (null != DbAsset)
-                    _dbContext.DeleteAsync<AssetDb>(DbAsset.Id).GetAwaiter().GetResult();
-
                 _disposed = true;
             }
         }
@@ -47,7 +43,7 @@ namespace AssetInformationListener.Tests.E2ETests.Fixtures
 
         private AssetDb SaveAssetDb(AssetDb dbEntity)
         {
-            _dbContext.SaveAsync<AssetDb>(dbEntity).GetAwaiter().GetResult();
+            _dbFixture.SaveEntityAsync<AssetDb>(dbEntity).GetAwaiter().GetResult();
             dbEntity.VersionNumber = 0;
             return dbEntity;
         }
