+++ AssetInformationListener.Tests/E2ETests/Stories/AccountCreatedUpdatesPersonTenureTests.cs
@@ -1,5 +1,6 @@
 using AssetInformationListener.Tests.E2ETests.Fixtures;
 using AssetInformationListener.Tests.E2ETests.Steps;
+using Hackney.Core.Testing.DynamoDb;
 using System;
 using TestStack.BDDfy;
 using Xunit;
@@ -10,21 +11,21 @@ namespace AssetInformationListener.Tests.E2ETests.Stories
         AsA = "SQS Entity Listener",
         IWant = "a function to process the AccountCreated message",
         SoThat = "The correct details are set on the appropriate asset")]
-    [Collection("Aws collection")]
+    [Collection("AppTest collection")]
     public class AccountCreatedUpdatesPersonTenureTests : IDisposable
     {
-        private readonly AwsIntegrationTests _dbFixture;
+        private readonly IDynamoDbFixture _dbFixture;
         private readonly AssetFixture _assetFixture;
         private readonly TenureApiFixture _tenureApiFixture;
         private readonly AccountApiFixture _accountApiFixture;
 
         private readonly AccountCreatedSteps _steps;
 
-        public AccountCreatedUpdatesPersonTenureTests(AwsIntegrationTests dbFixture)
+        public AccountCreatedUpdatesPersonTenureTests(MockApplicationFactory appFactory)
         {
-            _dbFixture = dbFixture;
+            _dbFixture = appFactory.DynamoDbFixture;
 
-            _assetFixture = new AssetFixture(_dbFixture.DynamoDbContext);
+            _assetFixture = new AssetFixture(_dbFixture);
             _tenureApiFixture = new TenureApiFixture();
             _accountApiFixture = new AccountApiFixture();
