+++ AssetInformationListener.Tests/E2ETests/Stories/TenureCreatedOrUpdatedTests.cs
@@ -1,5 +1,6 @@
 using AssetInformationListener.Tests.E2ETests.Fixtures;
 using AssetInformationListener.Tests.E2ETests.Steps;
+using Hackney.Core.Testing.DynamoDb;
 using System;
 using TestStack.BDDfy;
 using Xunit;
@@ -10,20 +11,20 @@ namespace AssetInformationListener.Tests.E2ETests.Stories
         AsA = "SQS Entity Listener",
         IWant = "a function to process the TenureCreated message",
         SoThat = "the asset is updated with correct details fromn the new tenure")]
-    [Collection("Aws collection")]
+    [Collection("AppTest collection")]
     public class TenureCreatedOrUpdatedTests : IDisposable
     {
-        private readonly AwsIntegrationTests _dbFixture;
+        private readonly IDynamoDbFixture _dbFixture;
         private readonly AssetFixture _assetFixture;
         private readonly TenureApiFixture _tenureApiFixture;
 
         private readonly TenureCreatedOrUpdatedSteps _steps;
 
-        public TenureCreatedOrUpdatedTests(AwsIntegrationTests dbFixture)
+        public TenureCreatedOrUpdatedTests(MockApplicationFactory appFactory)
         {
-            _dbFixture = dbFixture;
+            _dbFixture = appFactory.DynamoDbFixture;
 
-            _assetFixture = new AssetFixture(_dbFixture.DynamoDbContext);
+            _assetFixture = new AssetFixture(_dbFixture);
             _tenureApiFixture = new TenureApiFixture();
 
             _steps = new TenureCreatedOrUpdatedSteps();
