+++ AssetInformationListener/AssetInformationListener.csproj
@@ -19,9 +19,9 @@
     <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="2.1.0" />
     <PackageReference Include="Amazon.Lambda.SQSEvents" Version="2.0.0" />
     <PackageReference Include="AWSXRayRecorder.Handlers.AwsSdk" Version="2.8.3" />
-    <PackageReference Include="Hackney.Core.DynamoDb" Version="1.30.0" />
-    <PackageReference Include="Hackney.Core.Logging" Version="1.30.0" />
-    <PackageReference Include="Hackney.Core.Sns" Version="1.38.0" />
+    <PackageReference Include="Hackney.Core.DynamoDb" Version="1.51.0" />
+    <PackageReference Include="Hackney.Core.Logging" Version="1.49.0" />
+    <PackageReference Include="Hackney.Core.Sns" Version="1.52.0" />
     <PackageReference Include="Hackney.Shared.Asset" Version="0.6.0" />
     <PackageReference Include="Hackney.Shared.Tenure" Version="0.9.0" />
     <PackageReference Include="Microsoft.CodeAnalysis.NetAnalyzers" Version="5.0.3">
